Create a minimal TAP-compatible testing infrastructure for c
We're going to be testing the interaction between a C++ infrastructure
and C data structures. We'll want to ensure that the C code works before
starting to diagnose the C++.

My favorite C++ testing tools _will_ work on C code, but only if you
compile with the C++ compiler. I want to ensure that the C is _pure_ C,
so we need something that works under the C compiler. I was deep into
this before I found taplib, which I may try in the future, but for now
this is good enough.