A really minimal linked-list stack implementation in C
I've used this kind of thing in code-gold a lot because there isn't much
to it, but the differences in the signatures (stack* vs stack**) always
annoys me.

Still, it explifies what I think of as the default implementation:
singly-linked, using "next" as the forward pointer, and NULL to mark the
end of the list.