A more sophisticated list implementation.
This one is a double-linked list and uses a sentinal node to establish a
starting (and ending) point for traversals. It needs a more
sophisticated infrastructure, but the good news is that the signatures
are more uniform.

Two big differences from the "default" implemetnation will test the
flexibility of the C++ infrastructure we're building: the marker for the
end of a traversal is no longer NULL, and reverse traversals need a
different name for the "forward" link (that is "prev").