Lift the view on opaque node types
For reason I haven't yet sussed out, the list node types need to be
complete to call being() and end(). This may be related to having an
explicit list type in those templates (coming soon to a commit near
you!), but I'm not sure how I can beat that yet.