Extend tests for dl_deque
Up to now, the only tests of the iterator interaction with the dl_deque
structure were for the creation and comparison of iterators.

Here we test walking the list with explicit iterator loops (both
directions), and using sstd::find_if similar to what is done for
c_stack. Could probably use an explicit test for using algorithms
backwards.